Invert ownership between PyFunction and THPFunction.
Fixes #16532 and #14960.

This patch is a massive hack.  The way I constructed it was I flipped the
ownership between PyFunction and THPFunction, but maintained a weak pointer from
THPFunction to PyFunction so all existing code works.  Essentially, this patch
assumes that PyFunction stays live as long as you have a THPFunction: intuitively,
this makes sense, since the ctx object should only really stay live as long as
you're actually going to execute the backwards, which will keep the PyFunction
live.  But as you can see from the presently skipped tests (specifically,
test_hook_none), this is not always true.  But it seems to be true for the code
we care about, and that's enough for me!

Some subtleties:
- PyFunction is a C++ object that refers to a PyObject.  This means it needs
  a custom deleter to handle deleting the PyObject, since you can't assume
  you have the GIL when it dies.
- The old test_gc_in_destructor failed our internal assert because we never
  actually ran a backwards, and thus never actually materialized PyFunction.
  I'm chalking this up as "misuse of API" and rewrote the test to not have
  this problem.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: a25e8404551330fb7c6903274c1b9bc2faf4e2a2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/22983